Multi-cored process changes
Changes how yielding works and made transitions to and from _M states
more explicit.  Vcore0 can yield and come back up at the hart entry (you
need to prepare_for_multi_mode()).  In general, all cores are equal when
in an _M state, and vcore0 continues running the calling context only on
transition from _S to _M, and not whenever proc_run()ing a RUNNABLE_M.

Added documentation about how processes work, mostly with respect to
the kernel interface.  Part 2 has specific info about the transitions to
and from _M.

Finally moved current to per_cpu_info, and also added current_tf.
Current is still a pointer to the process that is on the core.
Current_tf is a pointer to the trapframe that is on the core's kernel
stack.  This is also stored in per_cpu_info.  It could be discovered
through arcane hackery, but probably isn't worth it.

+++ kern/arch/i386/smp_boot.c
@@ -28,7 +28,7 @@
 #include <timing.h>
 
 extern handler_wrapper_t (RO handler_wrappers)[NUM_HANDLER_WRAPPERS];
-volatile uint8_t num_cpus = 0xee;
+volatile uint32_t num_cpus = 0xee;
 uintptr_t RO smp_stack_top;
 
 #define DECLARE_HANDLER_CHECKLISTS(vector)                          \
